+++ lib/src/cancelablethen.dart
@@ -1,5 +1,5 @@
-///Canncelable a callback function for Future.
-///Ofcouse, we can not cancel a Future, this is just ignore call callback [then]
+///Cancelable a callback function for Future.
+///Of course, we can not cancel a Future, this is just ignore call callback [then]
 ///
 ///Future return null if it's canceled
 class CancelableThen<T> {